Refactor common test server functions
